........
  r683758 | josh | 2008-08-08 00:49:10 +0100 (Fri, 08 Aug 2008) | 1 line
  
  Consolidating ValueRecordsAggregate within RowRecordsAggregate
........
  r683788 | josh | 2008-08-08 02:30:30 +0100 (Fri, 08 Aug 2008) | 1 line
  
  Converted RowRecordsAggregate to proper RecordAggregate
........
  r683871 | josh | 2008-08-08 07:27:06 +0100 (Fri, 08 Aug 2008) | 1 line
  
  Extracting PageSettingsBlock from Sheet
........
  r683880 | josh | 2008-08-08 07:56:06 +0100 (Fri, 08 Aug 2008) | 1 line
  
  Finished extracting PageSettingsBlock logic from Sheet
........
  r683901 | josh | 2008-08-08 09:05:07 +0100 (Fri, 08 Aug 2008) | 1 line
  
  Patch 45577 - Added implementations for Excel functions NOW and TODAY, added property getters to HSSFConditionalFormattingRule
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@684213 13f79535-47bb-0310-9956-ffa450edef68

+++ src/ooxml/interfaces-jdk15/org/apache/poi/ss/usermodel/Row.java
@@ -59,8 +59,7 @@ public interface Row extends Iterable<Cell> {
      *
      * @return HSSFCell a high level representation of the created cell.
      */
-
-    Cell createCell(short column, int type);
+    Cell createCell(int column, int type);
 
     /**
      * remove the HSSFCell from this row.
